auth_service/requirements.txt
Flask==2.2.2
gevent==22.10.1
flask-redoc==0.2.1
flask_sqlalchemy==3.0.3
redis==4.5.1
flask_jwt_extended==4.4.4
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.5
aioredis==2.0.1
flasgger==0.9.5
apispec==6.0.2
marshmallow==3.19.0
pydantic==1.10.5
alembic==1.9.4

